Move shared contexts under contexts folder
