+++ app/src/editor/canvas/components/CanvasController/useCanvas.jsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 
-import { Context as UndoContext } from '$shared/components/UndoContextProvider'
+import { Context as UndoContext } from '$shared/contexts/Undo'
 
 export default function useCanvas() {
     const { state } = useContext(UndoContext)
