+++ app/src/editor/canvas/index.jsx
@@ -17,12 +17,12 @@ import routes from '$routes'
 
 import isEditableElement from '$editor/shared/utils/isEditableElement'
 import UndoControls from '$editor/shared/components/UndoControls'
-import * as UndoContext from '$shared/components/UndoContextProvider'
-import { Provider as PendingProvider } from '$shared/components/PendingContextProvider'
-import Subscription from '$shared/components/SubscriptionStatusContextProvider/Subscription'
-import * as SubscriptionStatus from '$shared/components/SubscriptionStatusContextProvider'
-import { ClientProvider } from '$shared/components/StreamrClientContextProvider'
-import { Provider as ModalProvider } from '$shared/components/ModalContextProvider'
+import * as UndoContext from '$shared/contexts/Undo'
+import { Provider as PendingProvider } from '$shared/contexts/Pending'
+import Subscription from '$shared/components/Subscription'
+import * as SubscriptionStatus from '$shared/contexts/SubscriptionStatus'
+import { Provider as ClientProvider } from '$shared/contexts/StreamrClient'
+import { Provider as ModalProvider } from '$shared/contexts/ModalApi'
 import * as sharedServices from '$editor/shared/services'
 import BodyClass from '$shared/components/BodyClass'
 import Sidebar from '$editor/shared/components/Sidebar'
