+++ app/src/editor/dashboard/index.jsx
@@ -7,12 +7,12 @@ import withErrorBoundary from '$shared/utils/withErrorBoundary'
 import LoadingIndicator from '$userpages/components/LoadingIndicator'
 import ErrorComponentView from '$shared/components/ErrorComponentView'
 import UndoControls from '$editor/shared/components/UndoControls'
-import { Context as UndoContext, Provider as UndoContextProvider } from '$shared/components/UndoContextProvider'
-import { ClientProvider } from '$shared/components/StreamrClientContextProvider'
+import { Context as UndoContext, Provider as UndoContextProvider } from '$shared/contexts/Undo'
+import { Provider as ClientProvider } from '$shared/contexts/StreamrClient'
 import * as sharedServices from '$editor/shared/services'
-import { Provider as ModalProvider } from '$shared/components/ModalContextProvider'
+import { Provider as ModalProvider } from '$shared/contexts/ModalApi'
 import { SelectionProvider } from '$editor/shared/hooks/useSelection'
-import { Provider as PendingProvider } from '$shared/components/PendingContextProvider'
+import { Provider as PendingProvider } from '$shared/contexts/Pending'
 import { useAnyPending } from '$shared/hooks/usePending'
 import CanvasStyles from '$editor/canvas/index.pcss'
 import Sidebar from '$editor/shared/components/Sidebar'
