+++ app/src/marketplace/containers/EditProductPage/index.jsx
@@ -5,7 +5,7 @@ import { withRouter } from 'react-router-dom'
 import { I18n } from 'react-redux-i18n'
 
 import CoreLayout from '$shared/components/Layout/Core'
-import * as UndoContext from '$shared/components/UndoContextProvider'
+import * as UndoContext from '$shared/contexts/Undo'
 import Toolbar from '$shared/components/Toolbar'
 import type { Product } from '$mp/flowtype/product-types'
 import { isCommunityProduct } from '$mp/utils/product'
@@ -24,7 +24,7 @@ import BackButton from './BackButton'
 import Editor from './Editor'
 import Preview from './Preview'
 import ProductEditorDebug from './ProductEditorDebug'
-import { Provider as ModalProvider } from '$shared/components/ModalContextProvider'
+import { Provider as ModalProvider } from '$shared/contexts/ModalApi'
 import ConfirmSaveModal from './ConfirmSaveModal'
 import DeployCommunityModal from './DeployCommunityModal'
 import PublishModal from './PublishModal'
