+++ app/src/marketplace/containers/ProductController/useEditableProduct.js
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 
-import { Context as UndoContext } from '$shared/components/UndoContextProvider'
+import { Context as UndoContext } from '$shared/contexts/Undo'
 
 export default function useEditableProduct() {
     const { state } = useContext(UndoContext)
